# backend/openapi.yaml
openapi: 3.1.0
info:
  title: JAAKD Trading API
  version: 1.0.0
  description: |
    REST API for the JAAKD direct trading platform (Spring Boot service in `backend/`).

    - Every endpoint requires a Supabase-issued JWT in the `Authorization: Bearer <token>` header.
    - A user can only see and change their own data (BR-02): every query is filtered by the
      user ID taken from the JWT's `sub` claim.
    - All errors share one shape, `ErrorResponse` (see `GlobalExceptionHandler`).
    - Endpoints tagged **Orders** and the **instrument prices** endpoint are planned (contract first)
      and not implemented yet.
  license:
    name: Internal use only
servers:
  - url: http://localhost:8081
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/holdings/portfolio/{portfolioId}:
    get:
      tags: [Holdings]
      summary: List the holdings in one of the user's portfolios
      description: Returns an empty list if the portfolio does not exist or is not owned by the user.
      operationId: getHoldingsByPortfolio
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Holdings in the portfolio
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HoldingResponse"
              example:
                - holdingId: 3f6c1a2e-8b4d-4c1a-9e2f-1a2b3c4d5e6f
                  portfolioId: 9a8b7c6d-5e4f-4a3b-2c1d-0e9f8a7b6c5d
                  instrumentId: 1b2c3d4e-5f6a-4b7c-8d9e-0f1a2b3c4d5e
                  currentQuantity: 10
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/portfolios:
    get:
      tags: [Portfolios]
      summary: List the signed-in user's portfolios
      operationId: getCurrentUserPortfolios
      responses:
        "200":
          description: The user's portfolios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioResponse"
              example:
                - portfolioId: 9a8b7c6d-5e4f-4a3b-2c1d-0e9f8a7b6c5d
                  portfolioName: Long-term
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [Portfolios]
      summary: Create a portfolio for the signed-in user
      operationId: createPortfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioRequest"
            example:
              portfolioName: Long-term
      responses:
        "201":
          description: Portfolio created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioResponse"
              example:
                portfolioId: 9a8b7c6d-5e4f-4a3b-2c1d-0e9f8a7b6c5d
                portfolioName: Long-term
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/portfolios/{portfolioId}:
    get:
      tags: [Portfolios]
      summary: Get one of the user's portfolios
      operationId: getPortfolioById
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: The portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags: [Portfolios]
      summary: Rename a portfolio
      operationId: updatePortfolio
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePortfolioRequest"
            example:
              portfolioName: Retirement
      responses:
        "200":
          description: Portfolio updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [Portfolios]
      summary: Delete a portfolio (and its holdings and watchlist items)
      operationId: deletePortfolio
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "204":
          description: Portfolio deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  # ------------------------------------------------------------------ Orders
...
  /api/order-logs/portfolio/{portfolioId}:
    get:
      tags: [OrderLogs]
      summary: List a portfolio's order history, newest first (the blotter, BR-11)
      operationId: getOrderLogsByPortfolio
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Order log entries, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderLogResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/trades/holding/{holdingId}:
    get:
      tags: [Trades]
      summary: List the trades that changed a holding
      operationId: getTradesByHolding
      parameters:
        - $ref: "#/components/parameters/HoldingId"
      responses:
        "200":
          description: Trades for the holding
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/trades/{tradeId}:
    get:
      tags: [Trades]
      summary: Get one trade
      operationId: getTradeById
      parameters:
        - $ref: "#/components/parameters/TradeId"
      responses:
        "200":
          description: The trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResponse"
              example:
                tradeId: 2e3f4a5b-6c7d-4e8f-9a0b-1c2d3e4f5a6b
                holdingId: 3f6c1a2e-8b4d-4c1a-9e2f-1a2b3c4d5e6f
                orderLogId: 5d2a1f0e-3c4b-4a59-8e7d-6f5a4b3c2d1e
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/trades:
    post:
      tags: [Trades]
      summary: Record a trade for an order log entry
      description: >
        If `holdingId` is omitted, the holding for the order's portfolio + instrument is
        found, or created if it does not exist. Only one trade may exist per order log entry.
      operationId: createTrade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTradeRequest"
            example:
              orderLogId: 5d2a1f0e-3c4b-4a59-8e7d-6f5a4b3c2d1e
      responses:
        "201":
          description: Trade created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/ServerError"

  # -------------------------------------------------------------- Watchlists
  /api/watchlists/portfolio/{portfolioId}:
    get:
      tags: [Watchlists]
      summary: List the watchlist items in a portfolio
      operationId: getWatchlistItemsByPortfolio
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Watchlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItemResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/watchlists/{listItemId}:
    get:
      tags: [Watchlists]
      summary: Get one watchlist item
      operationId: getWatchlistItemById
      parameters:
        - $ref: "#/components/parameters/ListItemId"
      responses:
        "200":
          description: The watchlist item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItemResponse"
              example:
                listItemId: 4a5b6c7d-8e9f-4a0b-1c2d-3e4f5a6b7c8d
                portfolioId: 9a8b7c6d-5e4f-4a3b-2c1d-0e9f8a7b6c5d
                instrumentId: 1b2c3d4e-5f6a-4b7c-8d9e-0f1a2b3c4d5e
                name: Tech picks
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags: [Watchlists]
      summary: Rename a watchlist item
      operationId: updateWatchlistItem
      parameters:
        - $ref: "#/components/parameters/ListItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWatchlistItemRequest"
            example:
              name: Dividend picks
      responses:
        "200":
          description: Watchlist item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [Watchlists]
      summary: Remove an instrument from the watchlist
      operationId: deleteWatchlistItem
      parameters:
        - $ref: "#/components/parameters/ListItemId"
      responses:
        "204":
          description: Watchlist item deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/watchlists:
    post:
      tags: [Watchlists]
      summary: Add an instrument to a portfolio's watchlist
      operationId: createWatchlistItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWatchlistItemRequest"
            example:
              portfolioId: 9a8b7c6d-5e4f-4a3b-2c1d-0e9f8a7b6c5d
              instrumentId: 1b2c3d4e-5f6a-4b7c-8d9e-0f1a2b3c4d5e
              name: Tech picks
      responses:
        "201":
          description: Watchlist item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/ServerError"

  # ----------------------------------------------------------------- Profile
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token (ES256), validated against the project's JWKS
  parameters:
    HoldingId:
      name: holdingId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PortfolioId:
      name: portfolioId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    TradeId:
      name: tradeId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ListItemId:
      name: listItemId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Malformed request - failed validation (details lists each bad field)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example:
            timestamp: "2026-09-23T14:30:00Z"
            message: Validation failed
            details:
              currentQuantity: Current quantity must be greater than 0
    Unauthorized:
      description: Missing, expired or invalid JWT
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource does not exist, or is not owned by the signed-in user
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example:
            timestamp: "2026-09-23T14:30:00Z"
            message: Holding not found
            details: {}
    Conflict:
      description: Resource already exists (duplicate)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ServerError:
      description: Unexpected error - generic message only, no stack trace
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        message: { type: string }
        details:
          type: object
          additionalProperties: { type: string }

    # ---- enums
...
    OrderSide:
      type: string
      enum: [BUY, SELL, DEPOSIT, WITHDRAW]
    OrderStatus:
      type: string
      enum: [SUBMITTED, PENDING, CANCELLED, ACCEPTED, REJECTED, EXECUTED, FAILED]

    # ---- holdings
...
    HoldingResponse:
      type: object
      properties:
        holdingId: { type: string, format: uuid }
        portfolioId: { type: string, format: uuid }
        instrumentId: { type: string, format: uuid }
        currentQuantity: { type: number, format: double }

    # ---- instruments
...
    CreatePortfolioRequest:
      type: object
      required: [portfolioName]
      properties:
        portfolioName: { type: string, minLength: 1, maxLength: 100 }
    UpdatePortfolioRequest:
      type: object
      required: [portfolioName]
      properties:
        portfolioName: { type: string, minLength: 1, maxLength: 100 }
    PortfolioResponse:
      type: object
      properties:
        portfolioId: { type: string, format: uuid }
        portfolioName: { type: string }

    # ---- orders (planned)
...
    OrderLogResponse:
      type: object
      properties:
        logOrderId: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        portfolioId: { type: string, format: uuid }
        instrumentId: { type: string, format: uuid }
        side: { $ref: "#/components/schemas/OrderSide" }
        quantity: { type: number, format: double }
        timestamp: { type: string, format: date-time }
        metadata: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/OrderStatus" }
        executionPrice: { type: number, format: double }

    # ---- trades
    CreateTradeRequest:
      type: object
      required: [orderLogId]
      properties:
        holdingId:
          type: [string, "null"]
          format: uuid
          description: Optional - if omitted, the holding is found or created from the order log
        orderLogId: { type: string, format: uuid }
    TradeResponse:
      type: object
      properties:
        tradeId: { type: string, format: uuid }
        holdingId: { type: string, format: uuid }
        orderLogId: { type: string, format: uuid }

    # ---- watchlists
    CreateWatchlistItemRequest:
      type: object
      required: [portfolioId, instrumentId]
      properties:
        portfolioId: { type: string, format: uuid }
        instrumentId: { type: string, format: uuid }
        name: { type: [string, "null"], maxLength: 255 }
    UpdateWatchlistItemRequest:
      type: object
      properties:
        name: { type: [string, "null"], maxLength: 255 }
    WatchlistItemResponse:
      type: object
      properties:
        listItemId: { type: string, format: uuid }
        portfolioId: { type: string, format: uuid }
        instrumentId: { type: string, format: uuid }
        name: { type: [string, "null"] }

    # ---- profile
...